Fix exim.conf primary_hostname regexp to match line
The regexp should match the inserted line itself.
This fixes: "Error: Validation of File_line[exim-set-primary-hostname]
failed: When providing a 'match' parameter, the value must be a
regex that matches against the value of your 'line' parameter"

It also prevents file_line appending the line to the bottom of the file
which could cause a syntax error.