fix!: allow pem files to be specified
This allows to specify pem files which are used to create combined
certificates. These certificates are used by DirectAdmin in the
Exim configuration if you're using SNI. Adds tests, too.

BREAKING CHANGE: Instead of assuming the variables were used to
specify a file we now assume that the actual content is in the
parameters sslcert, sslkey, sslca and sslpem.

fixes #15